Move to the Master branch
I am wondering if changing directly to the master/main branch in this manner is ok since we can do it whenever we desire to keep this package updated. No need to use commits. Simply use the main branch and from time to time easily change the shasum and the ynh versioning number and we are done. RSS Bridge does not have releases anymore. And this ynh package is not updated in a year or so.